Refactoring and other fixes
