+++ hw2/.vscode/launch.json
@@ -1,0 +1,42 @@
+{
+    // Use IntelliSense to learn about possible attributes.
+    // Hover to view descriptions of existing attributes.
+    // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+    "version": "0.2.0",
+    "configurations": [
+        {
+            "type": "java",
+            "name": "Current File",
+            "request": "launch",
+            "mainClass": "${file}"
+        },
+        {
+            "type": "java",
+            "name": "Main",
+            "request": "launch",
+            "mainClass": "com.solvd.hw2.Main",
+            "projectName": "solvd"
+        },
+        {
+            "type": "java",
+            "name": "JaxbMain",
+            "request": "launch",
+            "mainClass": "com.solvd.hw2.jaxb.JaxbMain",
+            "projectName": "solvd"
+        },
+        {
+            "type": "java",
+            "name": "MyBatisMain",
+            "request": "launch",
+            "mainClass": "com.solvd.hw2.mybatis.MyBatisMain",
+            "projectName": "solvd"
+        },
+        {
+            "type": "java",
+            "name": "ParserMain",
+            "request": "launch",
+            "mainClass": "com.solvd.hw2.parsers.ParserMain",
+            "projectName": "solvd"
+        }
+    ]
+}
