+++ hw2/src/main/java/com/solvd/hw2/dao/jdbc/ClientTypeDao.java
@@ -1,4 +1,4 @@
-package com.solvd.hw2.dao;
+package com.solvd.hw2.dao.jdbc;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -6,24 +6,23 @@ import java.util.ArrayList;
 import java.util.List;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-
 import com.solvd.hw2.CustomPool;
 import com.solvd.hw2.dao.abstracts.Dao;
-import com.solvd.hw2.models.Institution;
+import com.solvd.hw2.models.ClientType;
 
-public class InstitutionsDao extends Dao
+public class ClientTypeDao extends Dao
 {
-    private static final Logger LOGGER = LogManager.getLogger("Institution DAO");
-    private static final String INSTITUTION_TABLE = "institutions";
-    private static final String ID_COL = "institutionId";
-    private static final String NAME_COL = "intstitutionName";
+    private static final Logger LOGGER = LogManager.getLogger("Client Type DAO");
+    private static final String CLIENT_TYPE_TABLE = "clientTypes";
+    private static final String ID_COL = "clientTypeId";
+    private static final String NAME_COL = "clientTypeName";
 
-    public List<Institution> select(ArrayList<String> fields, Institution criteriaVals, String operator)
+    public List<ClientType> select(ArrayList<String> fields, ClientType criteriaVals, String operator)
     {
         try
         {
-            ArrayList<Institution> ret = new ArrayList<Institution>();
-            ResultSet results = getSelectResults(fields, criteriaVals, INSTITUTION_TABLE, operator);
+            ArrayList<ClientType> ret = new ArrayList<ClientType>();
+            ResultSet results = getSelectResults(fields, criteriaVals, CLIENT_TYPE_TABLE, operator);
 
             while (results.next())
             {
@@ -43,10 +42,10 @@ public class InstitutionsDao extends Dao
                     }
                 }
 
-                ret.add(new Institution(newId, newName));
+                ret.add(new ClientType(newId, newName));
             }
             
-            CustomPool.releaseConn();
+            CustomPool.releaseConn();           
             return ret;
         }
 
@@ -56,4 +55,5 @@ public class InstitutionsDao extends Dao
             return null;
         }
     }
+
 }
