+++ hw2/src/main/java/com/solvd/hw2/dao/jdbc/InstitutionsDao.java
@@ -1,37 +1,33 @@
-package com.solvd.hw2.dao;
+package com.solvd.hw2.dao.jdbc;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
-
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-
 import com.solvd.hw2.CustomPool;
 import com.solvd.hw2.dao.abstracts.Dao;
-import com.solvd.hw2.models.Degree;
+import com.solvd.hw2.models.Institution;
 
-public class DegreeDao extends Dao
+public class InstitutionsDao extends Dao
 {
-    private static final Logger LOGGER = LogManager.getLogger("Degree DAO");
-    private final String DEGREE_TABLE = "degrees";
-    private final String ID_COL = "degreeId";
-    private final String NAME_COL = "name";
-    private final String YEAR_COL = "numOfYears";
+    private static final Logger LOGGER = LogManager.getLogger("Institution DAO");
+    private static final String INSTITUTION_TABLE = "institutions";
+    private static final String ID_COL = "institutionId";
+    private static final String NAME_COL = "intstitutionName";
 
-    public List<Degree> select(ArrayList<String> fields, Degree criteriaVals, String operator)
+    public List<Institution> select(ArrayList<String> fields, Institution criteriaVals, String operator)
     {
         try
         {
-            ArrayList<Degree> ret = new ArrayList<Degree>();
-            ResultSet results = getSelectResults(fields, criteriaVals, DEGREE_TABLE, operator);
+            ArrayList<Institution> ret = new ArrayList<Institution>();
+            ResultSet results = getSelectResults(fields, criteriaVals, INSTITUTION_TABLE, operator);
 
             while (results.next())
             {
                 Integer newId = null;
-                String name = null;
-                Integer years = null;
+                String newName = null;
 
                 for (int i = 1; i <= results.getMetaData().getColumnCount(); i++)
                 {
@@ -42,16 +38,11 @@ public class DegreeDao extends Dao
 
                     else if (results.getMetaData().getColumnLabel(i).equals(NAME_COL))
                     {
-                        name = results.getString(i);
-                    }
-
-                    else if (results.getMetaData().getColumnLabel(i).equals(YEAR_COL))
-                    {
-                        years = results.getInt(i);
+                        newName = results.getString(i);
                     }
                 }
 
-                ret.add(new Degree(newId, name, years));
+                ret.add(new Institution(newId, newName));
             }
             
             CustomPool.releaseConn();
