+++ hw2/src/main/java/com/solvd/hw2/dao/jdbc/LabDao.java
@@ -1,4 +1,4 @@
-package com.solvd.hw2.dao;
+package com.solvd.hw2.dao.jdbc;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -6,7 +6,6 @@ import java.util.ArrayList;
 import java.util.List;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-
 import com.solvd.hw2.CustomPool;
 import com.solvd.hw2.dao.abstracts.Dao;
 import com.solvd.hw2.models.*;
