+++ hw2/src/main/java/com/solvd/hw2/dao/jdbc/ReportDao.java
@@ -1,4 +1,4 @@
-package com.solvd.hw2.dao;
+package com.solvd.hw2.dao.jdbc;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -6,31 +6,30 @@ import java.util.ArrayList;
 import java.util.List;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-
 import com.solvd.hw2.CustomPool;
 import com.solvd.hw2.dao.abstracts.Dao;
-import com.solvd.hw2.models.*;
+import com.solvd.hw2.models.Report;
 
-public class LocationDao extends Dao
+public class ReportDao extends Dao
 {
-    private static final Logger LOGGER = LogManager.getLogger("Location DAO");
-    private static final String LOCATION_TABLE = "locations";
+    private static final Logger LOGGER = LogManager.getLogger("Report DAO");
+    private static final String REPORTS_TABLE = "reports";
     private static final String ID_COL = "locId";
-    private static final String CITY_COL = "city";
-    private static final String COUNTRY_COL = "country";
+    private static final String NAME_COL = "reportName";
+    private static final String LINK_COL = "link";
 
-    public List<Location> select(ArrayList<String> fields, Location criteriaVals, String operator)
+    public List<Report> select(ArrayList<String> fields, Report criteriaVals, String operator)
     {
         try
         {
-            ArrayList<Location> ret = new ArrayList<Location>();
-            ResultSet results = getSelectResults(fields, criteriaVals, LOCATION_TABLE, operator);
+            ArrayList<Report> ret = new ArrayList<Report>();
+            ResultSet results = getSelectResults(fields, criteriaVals, REPORTS_TABLE, operator);
 
             while (results.next())
             {
                 Integer newId = null;
-                String city = null;
-                String country = null;
+                String name = null;
+                String link = null;
 
                 for (int i = 1; i <= results.getMetaData().getColumnCount(); i++)
                 {
@@ -39,18 +38,18 @@ public class LocationDao extends Dao
                         newId = results.getInt(i);
                     }
 
-                    else if (results.getMetaData().getColumnLabel(i).equals(CITY_COL))
+                    else if (results.getMetaData().getColumnLabel(i).equals(NAME_COL))
                     {
-                        city = results.getString(i);
+                        name = results.getString(i);
                     }
 
-                    else if (results.getMetaData().getColumnLabel(i).equals(COUNTRY_COL))
+                    else if (results.getMetaData().getColumnLabel(i).equals(LINK_COL))
                     {
-                        country = results.getString(i);
+                        link = results.getString(i);
                     }
                 }
 
-                ret.add(new Location(newId, city, country));
+                ret.add(new Report(newId, name, link));
             }
             
             CustomPool.releaseConn();
