+++ hw2/src/main/java/com/solvd/hw2/dao/mybatis/ClientMBDao.java
@@ -1,4 +1,4 @@
-package com.solvd.hw2.mybatis.dao;
+package com.solvd.hw2.dao.mybatis;
 
 import java.util.List;
 import org.apache.ibatis.session.SqlSession;
