+++ hw2/src/main/java/com/solvd/hw2/dao/mybatis/ExperimentMBDao.java
@@ -1,4 +1,4 @@
-package com.solvd.hw2.mybatis.dao;
+package com.solvd.hw2.dao.mybatis;
 
 import org.apache.ibatis.session.SqlSession;
 import org.apache.ibatis.session.SqlSessionFactory;
@@ -44,6 +44,15 @@ public class ExperimentMBDao
         try(SqlSession session = factory.openSession())
         {
             return session.selectOne("ExperimentMap.getExperimentById", id);
+        }
+    }
+
+    public void updateExperiment(Experiment newVals)
+    {
+        try(SqlSession session = factory.openSession())
+        {
+            session.update("ExperimentMap.updateExperiment", newVals);
+            session.commit();
         }
     }
 }
