--- hw2/src/main/java/com/solvd/hw2/lambda/interfaces/IConcatArrayList.java
@@ -1,8 +1,0 @@
-package com.solvd.hw2.lambda.interfaces;
-
-import java.util.ArrayList;
-
-public interface IConcatArrayList<T> 
-{
-    String concatArrayList(ArrayList<T> list, String extraStuff);
-}
