+++ hw2/src/main/java/com/solvd/hw2/lambda/interfaces/IConcatList.java
@@ -1,0 +1,8 @@
+package com.solvd.hw2.lambda.interfaces;
+
+import java.util.List;
+
+public interface IConcatList<T> 
+{
+    String concatList(List<T> list, String extraStuff);
+}
