+++ hw2/src/main/java/com/solvd/hw2/models/ExperimentType.java
@@ -11,8 +11,8 @@ import com.solvd.hw2.models.abstracts.Model;
 @JsonRootName ("experimentType")
 public class ExperimentType extends Model
 {
-    private static final String ID_COL = "expTypeId";
-    private static final String TYPE_COL = "expTypeName";
+    private static final String ID_COL = "experimentTypeId";
+    private static final String TYPE_COL = "experimentTypeName";
 
     private Integer id;
     private String typeName;
