+++ hw2/src/main/java/com/solvd/hw2/services/abstracts/IExperimentService.java
@@ -1,0 +1,10 @@
+package com.solvd.hw2.services.abstracts;
+
+import com.solvd.hw2.models.Experiment;
+
+public interface IExperimentService 
+{
+    public void insertExperiment(Experiment toInsert);
+    public Experiment getExperimentById(int id);
+    public void updateExperiment(Experiment newVals);
+}
